添加配置文件
