+++ config/wn1_1.json
@@ -1,0 +1,27 @@
+{
+      "distance": 10.0,
+      "horizontal": 10.0,
+      "vertical": 20.0,
+      "name": "wn1_1",
+      "type": "countArrester",
+      "ROI": {
+          "x": 750,
+          "y":280,
+          "w": 280,
+          "h": 286
+      },
+      "startPoint": {
+          "x": 81,
+          "y": 60
+      },
+      "endPoint": {
+          "x": 81,
+          "y": 60
+      },
+      "centerPoint": {
+          "x": 127,
+          "y": 136
+      },
+      "startValue": 0.0,
+      "totalValue": 10.0
+  }
