+++ config/wn1_4.json
@@ -1,0 +1,27 @@
+{
+      "distance": 10.0,
+      "horizontal": 10.0,
+      "vertical": 20.0,
+      "name": "wn1_4",
+      "type": "digitPressure",
+      "ROI": {
+          "x": 648,
+          "y":928,
+          "w": 214,
+          "h": 276
+      },
+      "startPoint": {
+          "x": 0,
+          "y": 0
+      },
+      "endPoint": {
+          "x": 0,
+          "y": 0
+      },
+      "centerPoint": {
+          "x": 0,
+          "y": 0
+      },
+      "startValue": 0.0,
+      "totalValue": 0.0
+  }
