+++ config/wn2_1.json
@@ -1,0 +1,27 @@
+{
+      "distance": 10.0,
+      "horizontal": 10.0,
+      "vertical": 20.0,
+      "name": "wn2_1",
+      "type": "normalPressure",
+      "ROI": {
+          "x": 241,
+          "y": 257,
+          "w": 344,
+          "h": 325
+      },
+      "startPoint": {
+          "x": 69,
+          "y": 235
+      },
+      "endPoint": {
+          "x": 229,
+          "y": 56
+      },
+      "centerPoint": {
+          "x": 242,
+          "y": 225
+      },
+      "startValue": 0.0,
+      "totalValue": 200.0
+  }
